implement antimerdian-safe CRS transformations
- implement coordinate reference system (CRS) transformations that
correctly handle polygons that cross the antimeridian.

- transform_geodataframe() uses a reprojection toolbox that handles the
antimeridian correctly. pyVPRM.lib.reproject_fiona implements the
reprojection based on tools in the fiona package as described in the
geopandas documentation here:
https://geopandas.org/en/stable/docs/user_guide/reproject_fiona.html#re-projecting-using-gdal-with-rasterio-and-fiona

- This fiona method is underpinned by GDAL (https://gdal.org). It
correctly handles polygons that cross the antimeridian (180 degrees
longitude). geopandas.GeoDataFrame.to_crs() does not handle this
situation correctly.

- All of the MODIS sinusoidal
tiles (https://modis-land.gsfc.nasa.gov/MODLAND_grid.html) along the
edges of the sinusoidal projection will fail to produce a correct LCM
crop using the to_crs() method.

Also use "--ignore-unmapped" flag for ESMF_RegridWeightGen

Toward the "edges" of the projection the MODIS sinusoidal projection
results in MODIS tiles that relative to a longitude-latitude
rectangle (such as ERA5 tiles) appear skewed into long parallelograms
with a relatively small angle in two of the four corners. Because we are
using the MODIS tiles as the destination grid to regrid the meteorology
data this can result in an area of interest that is a sub-area within a
single ERA5 tile where there are two triangles on the east and west ends
of the ERA5 tile that do not map onto the MODIS tile we are regridding
to. Without the "--ignore-unmapped" flag ESMF_RegridWeightGen fails with
an error message in these cases.

Setting the --ignore-unmapped allows the regrid to proceed with NaNs
placed in the destination where there are no overlapping source pixels.